1.26/requirements.txt
# streamlit==1.26
streamlit-nightly
langchain
openai
duckduckgo-search